added npt test
