+++ tests/examples-without-external-providers/cloud-router-2-precision-time-npt/main.tf
@@ -3,27 +3,24 @@ provider "equinix" {
   client_secret = var.equinix_client_secret
 }
 
-module "create_port_2_precision_time_ntp_service_profile" {
-  source = "../../../modules/port-connection"
-
-  connection_name       = var.connection_name
-  connection_type       = var.connection_type
-  notifications_type    = var.notifications_type
-  notifications_emails  = var.notifications_emails
-  bandwidth             = var.bandwidth
+module "create_cloud_router_2_precision_time_ntp" {
+  source               = "../../../modules/cloud-router-connection"
+  connection_name      = var.connection_name
+  connection_type      = var.connection_type
+  notifications_type   = var.notifications_type
+  notifications_emails = var.notifications_emails
+  bandwidth            = var.bandwidth
   purchase_order_number = var.purchase_order_number
 
   # A-side
-  aside_port_name = var.aside_port_name
-  aside_vlan_tag  = var.aside_vlan_tag
+  aside_fcr_uuid = var.aside_fcr_uuid
 
   # Z-side
   zside_ap_type         = var.zside_ap_type
   zside_ap_profile_type = var.zside_ap_profile_type
   zside_location        = var.zside_location
-  zside_sp_name         = var.zside_sp_name
+  zside_fabric_sp_name  = var.zside_fabric_sp_name
 }
-
 
 resource "equinix_fabric_precision_time_service" "ntp" {
   type = "NTP"
@@ -33,7 +30,7 @@ resource "equinix_fabric_precision_time_service" "ntp" {
   }
   connections = [
     {
-      uuid = module.create_port_2_precision_time_ntp_service_profile.primary_connection_id
+      uuid = module.create_cloud_router_2_precision_time_ntp.primary_connection_id
     }
   ]
   ipv4 = {
