+++ tests/examples-without-external-providers/cloud-router-2-precision-time-npt/outputs.tf
@@ -1,5 +1,5 @@
 output "port_2_npt_connection_id" {
-  value = module.create_port_2_precision_time_ntp_service_profile.primary_connection_id
+  value = module.create_cloud_router_2_precision_time_ntp.primary_connection_id
 }
 
 output "ntp_ept_resource_id" {
