+++ tests/examples-without-external-providers/cloud-router-2-precision-time-npt/versions.tf
